requirements.txt
feedparser>=6.0
openai>=1.0
pyyaml>=6.0
python-dotenv>=1.0
jinja2>=3.1
apscheduler>=3.10
requests>=2.31
markdown>=3.5
pytest>=8.0
